Make a multivariate distribution through a copula

In other words, this function transforms a copula from its uniform scale to the original scale of the data, such as m3/s.

Usage

bind_marginals(d1, d2, cop)

Arguments

d1, d2

The first and second marginal distributions to bind.

cop

A copula object (has class "bicop").

Value

A bivariate probability distribution of class "bidst_cop"

Examples

d1 <- distionary::dst_norm(0, 1)
#> Error in loadNamespace(x): there is no package called ‘distionary’
d2 <- distionary::dst_exp(1)
#> Error in loadNamespace(x): there is no package called ‘distionary’
bind_marginals(
  d1, d2,
  cop_rvinecopulib("gumbel", rotation = 0, parameters = 2)
)
#> Error: object 'd1' not found
bind_marginals(
  d1, d2,
  cop_ig(2, 3)
)
#> Error: object 'd1' not found