Copula Objects

Currently, this function makes a copula object with nothing in it. The idea is that you can then set things downstream, with functions such as set_cdf() and set_mean(). The idea behind this function is expected to be in flux.

Usage

is_copula(object)

is.copula(object)

copula(variable = c("unknown", "continuous", "discrete", "mixed"))

Arguments

object

Object to be tested

variable

Type of random vector: "continuous", "discrete", "mixed", or "unknown".

Value

A copula object with nothing in it.